Fixes #25 - created the DataType folder and moved the appropriate examples into it.
